pyproject.toml
[project]
name = "desktop-pet"
version = "1.0.0"
description = "A cute desktop pet application with animations and rest reminders"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Desktop Pet Contributors"}
]
keywords = ["desktop", "pet", "pyqt", "animation", "reminder", "desktop-pet", "virtual-pet"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Desktop Environment",
    "Topic :: Games/Entertainment",
]
dependencies = [
    "PyQt6>=6.6.0",
    "Pillow>=10.0.0",
    "aiohttp>=3.9.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/desktop-pet"
Repository = "https://github.com/yourusername/desktop-pet"
Issues = "https://github.com/yourusername/desktop-pet/issues"
Documentation = "https://github.com/yourusername/desktop-pet#readme"

[project.scripts]
desktop-pet = "desktop_pet.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/desktop_pet"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/assets",
    "/config",
    "/pets/default",
]

[dependency-groups]
dev = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pyinstaller>=6.0.0",
]